[IMP] account_chart_update: Tax group compatibility
Related to https://github.com/OCA/account-financial-tools/pull/2123#discussion_r2209443251

It is very important that the order is correct, i.e., you must try to create (for example)
the accounts first, which will then be linked to the tax that will be created, or that error will occur.

Similarly, it is also necessary to take tax groups into account and process them before taxes
to prevent taxes from being created/updated for the wrong group.